OKTA-95428 Add the text that users with '?' can only be fetched by id.
We tried adding support for handling '?' in the path but ran into
a lot of issues, including causing a regression for Adobe. Considering
the low usage of this pattern, it was decided to treat it like we do
with the '/' character, for which you are supposed to get the data
via id.

More details at - https://oktawiki.atlassian.net/wiki/pages/viewpage.action?pageId=158204501